+++ poisson_approval/profiles/ProfileNoisyDiscrete.py
@@ -4,7 +4,7 @@ from poisson_approval.constants.constants import *
 from poisson_approval.constants.EquilibriumStatus import EquilibriumStatus
 from poisson_approval.containers.AnalyzedStrategies import AnalyzedStrategies
 from poisson_approval.profiles.ProfileCardinal import ProfileCardinal
-from poisson_approval.strategies.StrategyThreshold import StrategyThreshold
+from poisson_approval.strategies.StrategyThresholdOptimistic import StrategyThresholdOptimistic
 from poisson_approval.utils.DictPrintingInOrderIgnoringZeros import DictPrintingInOrderIgnoringZeros
 from poisson_approval.utils.Util import product_dict, sort_weak_order, is_weak_order, my_division
 from poisson_approval.utils.UtilCache import cached_property
@@ -378,7 +378,7 @@ d_weak_order_share={'a~b>c': Fraction(53, 100)})
 
         Yields
         ------
-        StrategyThreshold
+        StrategyThresholdOptimistic
             All possible group strategies of the profile.
         """
         def possible_thresholds(ranking):
@@ -391,7 +391,7 @@ d_weak_order_share={'a~b>c': Fraction(53, 100)})
         d_ranking_possible_thresholds = {ranking: possible_thresholds(ranking) for ranking in RANKINGS}
 
         for d_ranking_threshold in product_dict(d_ranking_possible_thresholds):
-            yield StrategyThreshold(d_ranking_threshold, profile=self)
+            yield StrategyThresholdOptimistic(d_ranking_threshold, profile=self)
 
     @classmethod
     def order_and_label(cls, t):